spelUtils
