+++ packages/core/modules/config/ctx.js
@@ -1,11 +1,14 @@
 import moment from "moment";
 import {
-  SqlString, sqlEmptyValue, spelEscape, spelFixList,
+  SqlString, sqlEmptyValue,
   stringifyForDisplay, wrapWithBrackets,
 } from "../utils/export";
 import {
   mongoEmptyValue, mongoFormatOp1, mongoFormatOp2,
 } from "../utils/mongoUtils";
+import {
+  spelFixList, spelEscape,
+} from "../utils/spelUtils";
 import {escapeRegExp} from "../utils/stuff";
 import {getTitleInListValues} from "../utils/listValues";
