+++ packages/examples/src/demo_switch/spel_concat.tsx
@@ -7,45 +7,14 @@ import ReactSelect from "./select";
 /**
  * @deprecated
  */
-const jsonLogicFormatConcat = (parts: SpelConcatPart[]) => {
-  if (parts && Array.isArray(parts) && parts.length) {
-    return parts
-      .map(part => part?.value ?? part)
-      .filter(r => r != undefined);
-  } else {
-    return undefined;
-  }
-};
-
-/**
- * @deprecated
- */
-const jsonLogicImportConcat = (val: any): SpelConcatPart[] => {
-  if (val == undefined)
-    return undefined;
-  const errors: string[] = [];
-  const parts = Array.isArray(val) ? val : [val];
-  const res = parts.filter(v => v != undefined).map(v => {
-    return {
-      type: "property", 
-      value: val as string
-    } as SpelConcatPart;
-  });
-  if (errors.length) {
-    throw new Error(errors.join("\n"));
-  }
-  return res;
-};
-
-/**
- * @deprecated
- */
 export const getCaseValueWidgetConfig = (config: BasicConfig): CaseValueWidget => ({
   ...config.widgets.case_value,
-  spelFormatValue: QbUtils.ExportUtils.spelFormatConcat,
-  spelImportValue: QbUtils.ExportUtils.spelImportConcat,
-  jsonLogic: jsonLogicFormatConcat,
-  jsonLogicImport: jsonLogicImportConcat,
+  spelFormatValue: QbUtils.SpelUtils.spelFormatConcat,
+  spelImportValue: QbUtils.SpelUtils.spelImportConcat,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
+  jsonLogic: (QbUtils.JsonLogicUtils as any).jsonLogicFormatConcat,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
+  jsonLogicImport: (QbUtils.JsonLogicUtils as any).jsonLogicImportConcat,
   factory: ({value, setValue, id, fieldDefinition}: WidgetProps) => 
     <ReactSelect 
       value={value as SpelConcatPart[]}
